Adding placeholder text to "JMSText" textarea
Currently, the "Enter some text here for the message body..." text appears inside the body of the JMSText <textarea> element. This is fairly inconvenient, as the user is required to delete this text before keying in custom text message. As such, the intention of the said text is akin to directing users "what to do". This would, therefore, be better served as a helpful placeholder for the <textarea>.